barbot_controller/pyproject.toml
[tool.poetry]
name = "barbot_controller"
version = "0.1.0"
description = ""
authors = ["Cliff Li <cliff.li@netlight.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"
flask = "^2.3.2"
flask-cors = "^3.0.10"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
